+++ inst/lib/tutorial/tutorial.js
@@ -848,9 +848,14 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
 
   this.$forEachExercise(function(exercise) {
 
+    // get the knitr options script block and detach it (will move to input div)
+    var options_script = exercise.children('script[data-ui-opts="1"]').detach();
+
+    var chunk_options = options_script.length == 1 ? JSON.parse(options_script.text()) : {};
+
     // capture label and caption
     var label = exercise.attr('data-label');
-    var caption = exercise.attr('data-caption');
+    var caption = chunk_options.caption;
 
     // helper to create an id
     function create_id(suffix) {
@@ -882,9 +887,6 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
     for (var i=lines; i<thiz.kMinLines;i++)
       code = code + "\n";
 
-    // get the knitr options script block and detach it (will move to input div)
-    var options_script = exercise.children('script[data-ui-opts="1"]').detach();
-
     // wrap the remaining elements in an output frame div
     exercise.wrapInner('<div class="tutorial-exercise-output-frame"></div>');
     var output_frame = exercise.children('.tutorial-exercise-output-frame');
@@ -895,8 +897,13 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
 
     // creating heading
     var panel_heading = $('<div class="panel-heading tutorial-panel-heading"></div>');
-    panel_heading.text(caption);
     input_div.append(panel_heading);
+    var panel_heading_left = $('<div class="tutorial-panel-heading-left"></div>');
+    var panel_heading_right = $('<div class="tutorial-panel-heading-right"></div>');
+    panel_heading.append(panel_heading_left);
+    panel_heading.append(panel_heading_right);
+
+    panel_heading_left.html(caption);
 
     // create body
     var panel_body = $('<div class="panel-body"></div>');
@@ -904,8 +911,7 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
 
     // function to add a submit button
     function add_submit_button(icon, style, text, check) {
-      var button = $('<a class="btn ' + style + ' btn-xs btn-tutorial-run ' +
-                       'pull-right"></a>');
+      var button = $('<a class="btn ' + style + ' btn-xs btn-tutorial-run"></a>');
       button.append($('<i class="fa ' + icon + '"></i>'));
       button.attr('type', 'button');
       button.append(' ' + text);
@@ -921,17 +927,16 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
         thiz.$removeSolution(exercise);
         thiz.$showExerciseProgress(label, button, true);
       });
-      panel_heading.append(button);
+      panel_heading_right.append(button);
       return button;
     }
 
-    var chunk_options = options_script.length == 1 ? JSON.parse(options_script.text()) : {};
+    // create run button
+    var run_button = add_submit_button("fa-play", "btn-success", "Run Code", false);
+
     // create submit answer button if checks are enabled
     if (chunk_options.has_checker)
       add_submit_button("fa-check-square-o", "btn-primary", "Submit Answer", true);
-
-    // create run button
-    var run_button = add_submit_button("fa-play", "btn-success", "Run Code", false);
 
     // create code div and add it to the input div
     var code_div = $('<div class="tutorial-exercise-code-editor"></div>');
@@ -1023,7 +1028,7 @@ Tutorial.prototype.$initializeExerciseEditors = function() {
     editor.getSession().on('change', updateAceHeight);
 
     // add hint/solution/startover buttons if necessary
-    thiz.$addSolution(exercise, panel_heading, editor);
+    thiz.$addSolution(exercise, panel_heading_left, editor);
 
     exercise.parents('.section').on('shown', function() {
       editor.resize(true);
@@ -1070,10 +1075,11 @@ Tutorial.prototype.$addSolution = function(exercise, panel_heading, editor) {
   var hintDiv = thiz.$exerciseHintDiv(label);
 
   // function to add a helper button
-  function addHelperButton(icon, caption) {
+  function addHelperButton(icon, caption, ele_class) {
     var button = $('<a class="btn btn-light btn-xs btn-tutorial-solution"></a>');
     button.attr('role', 'button');
     button.attr('title', caption);
+    button.addClass(ele_class);
     button.append($('<i class="fa ' + icon + '"></i>'));
     button.append(' ' + caption);
     panel_heading.append(button);
@@ -1082,7 +1088,7 @@ Tutorial.prototype.$addSolution = function(exercise, panel_heading, editor) {
 
   // function to add a hint button
   function addHintButton(caption) {
-    return addHelperButton("fa-lightbulb-o", caption);
+    return addHelperButton("fa-lightbulb-o", caption, "btn-tutorial-hint");
   }
 
   // helper function to record solution/hint requests
@@ -1095,7 +1101,7 @@ Tutorial.prototype.$addSolution = function(exercise, panel_heading, editor) {
 
   // add a startover button
   if (editor.tutorial.startover_code !== null) {
-    var startOverButton = addHelperButton("fa-refresh", "Start Over");
+    var startOverButton = addHelperButton("fa-refresh", "Start Over", "btn-tutorial-start-over");
     startOverButton.on('click', function() {
       editor.setValue(editor.tutorial.startover_code, -1);
       thiz.$clearExerciseOutput(exercise);